refactor code

+++ app/Http/Controllers/UserController.php
@@ -14,13 +14,13 @@ class UserController extends Controller
 {
     public function user()
     {
-        $users = User::with(['role', 'user_access_file', 'user_access_file.jurusan', 'user_access_file.prodi'])->latest()->paginate(8);
+        $users = User::with(['role', 'user_access_file', 'user_access_file.jurusan', 'user_access_file.prodi'])->latest('updated_at')->paginate(8);
         return view('user.home', compact('users'));
     }
 
     public function user_filter($role)
     {
-        $users = User::where('role_id', $role)->latest()->paginate(8);
+        $users = User::where('role_id', $role)->latest('updated_at')->paginate(8);
         return view('user.home', compact('users'));
     }
 
@@ -50,7 +50,7 @@ class UserController extends Controller
     {
         $request->validate([
             'name' => 'unique:users',
-            'email' => 'email|unique:users',
+            'email' => 'email:dns|unique:users',
             'password' => 'min:8|string',
             'confirm' => 'same:password',
         ], [
@@ -59,6 +59,19 @@ class UserController extends Controller
                 'password.min' => 'Password minimal 8 karakter!',
                 'confirm.same' => 'Password tidak sama, mohon periksa kembali password Anda.'
             ]);
+
+        if ($request->role_id == 2 && !$request->jurusan_id) {
+            return back()->with('error', 'Mohon pilih jurusan untuk Ketua Jurusan!');
+        } elseif ($request->role_id == 3 && !$request->jurusan_id) {
+            return back()->with('error', 'Mohon pilih program studi untuk Koorprodi!');
+        } elseif ($request->role_id == 4) {
+            $notAllNull = array_filter($request->prodi_id_auditor, function ($value) {
+                return $value != null;
+            });
+            if (count($notAllNull) == 0) {
+                return back()->with('error', 'Mohon pilih minimal 1 program studi untuk Auditor!');
+            }
+        }
 
         $jurusan_id = ($request->role_id == 3) ? Prodi::find($request->prodi_id)->jurusan_id : $request->jurusan_id;
         $user = User::create([
@@ -70,16 +83,18 @@ class UserController extends Controller
 
 
         if ($request->role_id == 4) {
-            foreach ($request->prodi_id_auditor as $prodi) {
-                if ($prodi) {
-                    UserAccessFile::create([
-                        'user_id' => $user->id,
-                        'jurusan_id' => $jurusan_id,
-                        'prodi_id' => $prodi
-                    ]);
-                }
+            $prodis_auditor = array_filter($request->prodi_id_auditor, function ($value) {
+                return $value != null;
+            });
+            $prodis_auditor = array_unique($prodis_auditor);
+            foreach ($prodis_auditor as $prodi) {
+                UserAccessFile::create([
+                    'user_id' => $user->id,
+                    'jurusan_id' => $jurusan_id,
+                    'prodi_id' => $prodi
+                ]);
             }
-        } else {
+        } elseif ($request->role_id == 2 || $request->role_id == 3) {
             UserAccessFile::create([
                 'user_id' => $user->id,
                 'jurusan_id' => $jurusan_id,
@@ -106,7 +121,7 @@ class UserController extends Controller
     public function change_user_action(Request $request, $id_user)
     {
         $user = User::find($id_user);
-        $rules = ['email' => 'email'];
+        $rules = ['email' => 'email:dns'];
 
         if ($request->name != $user->name) {
             $rules['name'] = 'unique:users';
@@ -123,49 +138,69 @@ class UserController extends Controller
 
         $access = UserAccessFile::where('user_id', $user->id)->get();
         if ($request->role_id == 3) {
-            $jurusan = Prodi::find($request->prodi_id)->jurusan_id;
-            if ($user->role_id == 2 || $user->role_id == 3 && $access[0]->prodi_id != $request->prodi_id) {
-                $access[0]->update([
-                    'prodi_id' => $request->prodi_id,
-                    'jurusan_id' => $jurusan
-                ]);
-            } elseif ($user->role_id == 1) {
-                UserAccessFile::create([
-                    'user_id' => $request->id,
-                    'prodi_id' => $request->prodi_id,
-                    'jurusan_id' => $jurusan
-                ]);
-            } else {
-                if ($access->count() > 1) {
-                    $access->where('id', '!=', $access[0]->id)->each(function ($userAccessFile) {
-                        $userAccessFile->delete();
-                    });
+            if ($request->prodi_id) {
+                $jurusan = Prodi::find($request->prodi_id)->jurusan_id;
+                if ($user->role_id == 2 || $user->role_id == 3 && $access[0]->prodi_id != $request->prodi_id) {
                     $access[0]->update([
                         'prodi_id' => $request->prodi_id,
                         'jurusan_id' => $jurusan
                     ]);
-                }
-            }
-        } elseif ($request->role_id == 4) {
-            if ($user->role_id == 2 || $user->role_id == 3) {
-                $access[0]->update([
-                    'prodi_id' => $request->prodi_id_auditor[0],
-                    'jurusan_id' => null
-                ]);
-                if (count($request->prodi_id_auditor) > 1) {
-                    foreach ($request->prodi_id_auditor as $prodi) {
-                        if ($prodi && $prodi != $access[0]->prodi_id) {
-                            UserAccessFile::create([
-                                'user_id' => $request->id,
-                                'prodi_id' => $prodi,
-                                'jurusan_id' => null
-                            ]);
-                        }
+                } elseif ($user->role_id == 1) {
+                    UserAccessFile::create([
+                        'user_id' => $request->id,
+                        'prodi_id' => $request->prodi_id,
+                        'jurusan_id' => $jurusan
+                    ]);
+                } else {
+                    if ($access->count() > 1) {
+                        $access->where('id', '!=', $access[0]->id)->each(function ($userAccessFile) {
+                            $userAccessFile->delete();
+                        });
+                        $access[0]->update([
+                            'prodi_id' => $request->prodi_id,
+                            'jurusan_id' => $jurusan
+                        ]);
                     }
                 }
-            } elseif ($user->role_id == 1) {
-                foreach ($request->prodi_id_auditor as $prodi) {
-                    if ($prodi) {
+            } else {
+                return back()->with('error', 'Mohon pilih program studi untuk Koorprodi!');
+            }
+        } elseif ($request->role_id == 4) {
+            $prodis_baru = array_map('intval', $request->prodi_id_auditor);
+            $notAllNull = array_filter($prodis_baru, function ($value) {
+                return $value != 0;
+            });
+            $notAllNull = array_unique($notAllNull);
+
+            if (count($notAllNull) != 0) {
+                if ($user->role_id == 2 || $user->role_id == 3) {
+                    $prodis_baru = array_values($notAllNull);
+                    $access[0]->update([
+                        'prodi_id' => $prodis_baru[0],
+                        'jurusan_id' => null
+                    ]);
+                    if (count($prodis_baru) > 1) {
+                        foreach ($prodis_baru as $prodi) {
+                            if ($prodi != $access[0]->prodi_id) {
+                                UserAccessFile::create([
+                                    'user_id' => $request->id,
+                                    'prodi_id' => $prodi,
+                                    'jurusan_id' => null
+                                ]);
+                            }
+                        }
+                    }
+                } elseif ($user->role_id == 1) {
+                    foreach ($notAllNull as $prodi) {
+                        UserAccessFile::create([
+                            'user_id' => $request->id,
+                            'prodi_id' => $prodi,
+                            'jurusan_id' => null
+                        ]);
+                    }
+                } else {
+                    UserAccessFile::where('user_id', $user->id)->delete();
+                    foreach ($notAllNull as $prodi) {
                         UserAccessFile::create([
                             'user_id' => $request->id,
                             'prodi_id' => $prodi,
@@ -174,53 +209,34 @@ class UserController extends Controller
                     }
                 }
             } else {
-                $prodis = UserAccessFile::where('user_id', $user->id)->pluck('prodi_id')->toArray();
-                $prodis_baru = array_map('intval', $request->prodi_id_auditor);
-                if (count($prodis) == count($prodis_baru)) {
-                    for ($i = 0; $i < count($prodis); $i++) {
-                        if ($prodis_baru[$i] && $prodis[$i] != $prodis_baru[$i]) {
-                            $access[$i]->update([
-                                'prodi_id' => $prodis_baru[$i]
-                            ]);
-                        } elseif ($prodis_baru[$i] == 0) {
-                            $access[$i]->delete();
-                        }
-                    }
-                } elseif (count($prodis) < count($prodis_baru)) {
-                    UserAccessFile::where('user_id', $user->id)->delete();
-                    foreach ($prodis_baru as $prodi) {
-                        if ($prodi) {
-                            UserAccessFile::create([
-                                'user_id' => $request->id,
-                                'prodi_id' => $prodi,
-                                'jurusan_id' => null
-                            ]);
-                        }
-                    }
-                }
+                return back()->with('error', 'Mohon pilih minimal 1 program studi untuk Auditor!');
             }
         } elseif ($request->role_id == 2) {
-            if ($user->role_id == 3 || $user->role_id == 2) {
-                $access[0]->update([
-                    'jurusan_id' => $request->jurusan_id,
-                    'prodi_id' => null
-                ]);
-            } elseif ($user->role_id == 4) {
-                if ($access->count() > 1) {
-                    $access->where('id', '!=', $access[0]->id)->each(function ($userAccessFile) {
-                        $userAccessFile->delete();
-                    });
+            if ($request->jurusan_id) {
+                if ($user->role_id == 3 || $user->role_id == 2) {
                     $access[0]->update([
+                        'jurusan_id' => $request->jurusan_id,
+                        'prodi_id' => null
+                    ]);
+                } elseif ($user->role_id == 4) {
+                    if ($access->count() > 1) {
+                        $access->where('id', '!=', $access[0]->id)->each(function ($userAccessFile) {
+                            $userAccessFile->delete();
+                        });
+                        $access[0]->update([
+                            'prodi_id' => null,
+                            'jurusan_id' => $request->jurusan_id
+                        ]);
+                    }
+                } else {
+                    UserAccessFile::create([
+                        'user_id' => $request->id,
                         'prodi_id' => null,
                         'jurusan_id' => $request->jurusan_id
                     ]);
                 }
             } else {
-                UserAccessFile::create([
-                    'user_id' => $request->id,
-                    'prodi_id' => null,
-                    'jurusan_id' => $request->jurusan_id
-                ]);
+                return back()->with('error', 'Mohon pilih jurusan untuk Ketua Jurusan!');
             }
         } elseif ($request->role_id == 1) {
             UserAccessFile::where('user_id', $user->id)->delete();
@@ -245,6 +261,7 @@ class UserController extends Controller
             ]);
         }
 
+        $user->touch();
         activity()
             ->performedOn($user)
             ->log('Mengubah data user dengan id ' . $user->id);
